+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.2.0 (Unreleased)
 
 * Add USO email prompt in error component.
+* Add slider animation control.
 
 
 # v0.1.1 (2021-06-23)